refactor: 각 클래스를 별도 파일로 분리
파일 분리와 함께 Game에 대한 빠진 테스트를 추가한다.
