+++ src/main/kotlin/car/racing/GameRecords.kt
@@ -1,0 +1,10 @@
+package car.racing
+
+class GameRecords {
+    private val _records = mutableListOf<List<Int>>()
+    val records: List<List<Int>>
+        get() = _records
+
+    fun record(cars: Cars): Boolean = _records.add(cars.getAllMoves())
+    fun reset(): Unit = _records.clear()
+}
