+++ src/test/kotlin/car/racing/CarRacingTest.kt
@@ -2,157 +2,9 @@ package car.racing
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FreeSpec
-import io.kotest.matchers.ints.shouldBeGreaterThanOrEqual
-import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 
 class CarRacingTest : FreeSpec({
-
-    "입력 화면" - {
-        fun fakeConsole(@Suppress("UNUSED_PARAMETER") msg: String) = Unit
-
-        "정상 입력으로 진행" - {
-            var fakeReadValue: Int? = 0
-            fun fakeReader(): String? = fakeReadValue?.toString()
-            val inputView = InputView(::fakeReader, ::fakeConsole)
-
-            "드라이버는 5명" {
-                fakeReadValue = 5
-                inputView.readDrivers() shouldBe fakeReadValue
-            }
-
-            "이동횟수는 3번" {
-                fakeReadValue = 3
-                inputView.readMoves() shouldBe fakeReadValue
-            }
-        }
-
-        "비정상 입력으로 진행" - {
-            var fakeReadValue: String? = null
-            fun fakeReader(): String? = fakeReadValue
-            val inputView = InputView(::fakeReader, ::fakeConsole)
-
-            fakeReadValue = "a"
-            "드라이버는 숫자가 아니기 때문에 IllegalArgumentException 발생" {
-                shouldThrowExactly<IllegalArgumentException> {
-                    inputView.readDrivers()
-                }
-            }
-
-            fakeReadValue = "error"
-            "이동횟수는 숫자가 아니기 때문에 IllegalArgumentException 발생" {
-                shouldThrowExactly<IllegalArgumentException> {
-                    inputView.readMoves()
-                }
-            }
-
-            fakeReadValue = null
-            "드라이버는 null 이기 때문에 IllegalArgumentException 발생" {
-                shouldThrowExactly<IllegalArgumentException> {
-                    inputView.readDrivers()
-                }
-            }
-
-            "이동횟수는 null 이기 때문에 IllegalArgumentException 발생" {
-                shouldThrowExactly<IllegalArgumentException> {
-                    inputView.readMoves()
-                }
-            }
-        }
-    }
-
-    "결과 화면" - {
-        var recordCount = 0
-        fun fakeConsole(msg: String) {
-            recordCount += 1
-        }
-
-        val resultView = ResultView(::fakeConsole)
-
-        "빈 경기결과는 1번 기록되어 있다" {
-            recordCount = 0
-            resultView.showRecords(GameRecords())
-            recordCount shouldBe 1
-        }
-
-        "유효한 경기는 1번 이상 기록되어 있다" {
-            recordCount = 0
-            val drivers = 3
-            val cars = Cars(drivers).apply { repeat(10) { runAll() } }
-            val records = GameRecords().apply { record(cars) }
-            resultView.showRecords(records)
-            recordCount shouldBeGreaterThanOrEqual 1
-        }
-    }
-
-    "자동차 상태는" - {
-        "좋은경우 앞으로 이동한다" {
-            CarCondition.GOOD.canGoForward() shouldBe true
-        }
-        "나쁜경우 앞으로 이동하지 못한다" {
-            CarCondition.TERRIBLE.canGoForward() shouldBe false
-        }
-    }
-
-    "자동차는" - {
-        "앞으로 이동할 수 있다" {
-            val car = Car()
-            repeat(10) {
-                car.run()
-            }
-            car.moves shouldBeGreaterThanOrEqual 0
-        }
-
-        "차량 상태가 좋은경우 시도횟수와 동일하게 앞으로 이동할 수 있다" {
-            val car = Car(CarCondition.GOOD)
-            val trying = 10
-            repeat(trying) {
-                car.run()
-            }
-            car.moves shouldBe trying
-        }
-        "차량 상태가 나쁜경우 앞으로 이동하지 못한다" {
-            val car = Car(CarCondition.TERRIBLE)
-            val trying = 10
-            repeat(trying) {
-                car.run()
-            }
-            car.moves shouldBe 0
-        }
-    }
-
-    "자동차들은" - {
-        "드라이버의 수만큼 이동거리를 알 수 있다" {
-            val drivers = 199
-            val cars = Cars(drivers = drivers)
-            cars.runAll()
-            cars.getAllMove().size shouldBe drivers
-        }
-        "한번도 이동하지 않으면 모두 이동거리가 0이다" {
-            val cars = Cars(drivers = 10)
-            cars.getAllMove().all { it == 0 } shouldBe true
-        }
-    }
-
-    "경기 기록은" - {
-        val drivers = 3
-        val cars = Cars(drivers)
-        val records = GameRecords()
-
-        records.reset()
-        "기록횟수를 알 수 있다" {
-            records.record(cars)
-            records.records.size shouldBe 1
-        }
-
-        records.reset()
-        "기록은 초기화 할 수 있다" {
-            records.record(cars)
-            records.reset()
-            records.records.size shouldBe 0
-        }
-    }
-
     "자동차 경주" - {
         val racing = CarRacing()
